Created spawn module and plugin for effect
